+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 #### Added
 - Add 'lazy_symbolication' config
   - [#2958](https://github.com/bpftrace/bpftrace/pull/2958)
+- Add ability to list all probes in a program
+  - [#2969](https://github.com/bpftrace/bpftrace/pull/2969)
 #### Changed
 #### Deprecated
 #### Removed